Fix monsters package init
