+++ monsters/__init__.py
@@ -1,4 +1,10 @@
-# monsters/__init__.py (このファイルも `monsters` フォルダ直下に置きます)
+# monsters package initialization
 
-from .monster_class import Monster, GROWTH_TYPE_AVERAGE, GROWTH_TYPE_EARLY, GROWTH_TYPE_LATE
+from .monster_class import (
+    Monster,
+    GROWTH_TYPE_AVERAGE,
+    GROWTH_TYPE_EARLY,
+    GROWTH_TYPE_LATE,
+)
 from .monster_data import ALL_MONSTERS
+
